+++ packages/untis/test/sort_test.dart
@@ -5,9 +5,7 @@ import '../../../testlib.dart';
 
 TestCase sortTestCase(List<List> input, List<int> expected) => () async {
       if (input.length != expected.length) throw 'Invalid test.';
-      final s = input
-          .map((e) => Substitution(e[0], e[1], '', '', false))
-          .toList()
+      final s = input.map((e) => Substitution(e[0], e[1], '', '')).toList()
         ..sort();
       for (var i = 0; i < s.length; i++) {
         expect(s[i].affectedClass, input[expected[i]][0]);