+++ packages/dsb/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0-alpha.4
+
+- Made the `Item.fromJson` better
+- Fixed `Session.getJsonString` having horrible code
+
 ## 0.1.0-alpha.3
 
 - Got rid of the multi-file architecture